backend/requirements.txt
fastapi
uvicorn
pandas
pdfplumber
python-docx
requests
io
pandas
openai
csv
pathlib
subprocess
tempfile
uuid
re
